+++ src/vue-router/redirect.js
@@ -1,16 +1,21 @@
 import { vueRoutes } from './routes'
+import { store, vuexTypes } from '@/vuex'
 
 const REDIRECT_TYPES = {
   email: 1,
+  purchasedAsset: 2,
 }
 
 export const resolveRedirect = (to, from, next) => {
   const encodedValue = to.path.replace('/r/', '')
-  const action = JSON.parse(atob(encodedValue))
+  const decodedValue = JSON.parse(atob(encodedValue))
 
-  switch (action.type) {
+  switch (decodedValue.type) {
     case REDIRECT_TYPES.email:
       handleEmailRedirect(encodedValue, next)
+      break
+    case REDIRECT_TYPES.purchasedAsset:
+      handlePurchasedAssetRedirect(decodedValue.meta, next)
       break
   }
 }
@@ -20,4 +25,27 @@ function handleEmailRedirect (encodedVerificationCode, next) {
     ...vueRoutes.login,
     params: { encodedVerificationCode },
   })
+}
+
+function handlePurchasedAssetRedirect (decodedPurchasedAssetInfo, next) {
+  const isLoggedIn = store.getters[vuexTypes.isLoggedIn]
+  const isPurchasedSuccessful = decodedPurchasedAssetInfo.is_success
+
+  if (isLoggedIn) {
+    next({
+      ...vueRoutes.assetsExplore,
+      params: { isAssetPurchased: isPurchasedSuccessful },
+    })
+  } else {
+    if (isPurchasedSuccessful) {
+      next({
+        ...vueRoutes.successfulPayment,
+        params: { purchasedAssetInfo: decodedPurchasedAssetInfo },
+      })
+    } else {
+      next({
+        ...vueRoutes.failedPayment,
+      })
+    }
+  }
 }