+++ src/vue/forms/SponsorshipForm.vue
@@ -1,87 +1,98 @@
 <template>
   <div class="sponsorship-form">
-    <form
-      novalidate
-      class="app__form"
-      @submit.prevent="isFormValid() && showConfirmation()"
-    >
-      <div class="app__form-row">
-        <div class="app__form-field">
-          <select-field
-            :value="form.assetCode"
-            @input="setAssetCode"
-            :label="'sponsorship-form.asset-lbl' | globalize"
-            name="sponsorship-form-asset-code"
-            @blur="touchField('form.assetCode')"
-            :disabled="formMixin.isDisabled"
-          >
-            <option
-              v-for="asset in balancesAssets"
-              :key="asset.code"
-              :value="asset.code"
+    <template v-if="balancesAssets.length">
+      <form
+        novalidate
+        class="app__form"
+        @submit.prevent="isFormValid() && showConfirmation()"
+      >
+        <div class="app__form-row">
+          <div class="app__form-field">
+            <select-field
+              :value="form.assetCode"
+              @input="setAssetCode"
+              :label="'sponsorship-form.asset-lbl' | globalize"
+              name="sponsorship-form-asset-code"
+              @blur="touchField('form.assetCode')"
+              :disabled="formMixin.isDisabled"
             >
-              {{ asset.name }}
-            </option>
-          </select-field>
+              <option
+                v-for="asset in balancesAssets"
+                :key="asset.code"
+                :value="asset.code"
+              >
+                {{ asset.name }}
+              </option>
+            </select-field>
+          </div>
         </div>
-      </div>
 
-      <div class="app__form-row">
-        <div class="app__form-field">
-          <amount-input-field
-            v-model="form.amount"
-            name="sponsorship-form-amount"
-            validation-type="outgoing"
-            :max="assetBalance.balance"
-            :label="'sponsorship-form.amount-lbl' | globalize"
-            :asset="form.assetCode"
-            :readonly="formMixin.isDisabled"
+        <div class="app__form-row">
+          <div class="app__form-field">
+            <amount-input-field
+              v-model="form.amount"
+              name="sponsorship-form-amount"
+              validation-type="outgoing"
+              :max="assetBalance.balance"
+              :label="'sponsorship-form.amount-lbl' | globalize"
+              :asset="form.assetCode"
+              :readonly="formMixin.isDisabled"
+            />
+          </div>
+        </div>
+
+        <div class="app__form-row">
+          <div class="app__form-field">
+            <readonly-field
+              :label="'sponsorship-form.total-amount' | globalize"
+              :value="totalAmount"
+              :error-message="getFieldErrorMessage(
+                'totalAmount',
+                {
+                  available: assetBalance.balance,
+                }
+              )"
+            />
+          </div>
+        </div>
+
+        <template v-if="formMixin.isConfirmationShown">
+          <form-confirmation
+            @cancel="hideConfirmation"
+            @ok="submit"
+            :is-pending="isFormSubmitting"
+            class="app__form-confirmation"
           />
-        </div>
-      </div>
+        </template>
 
-      <div class="app__form-row">
-        <div class="app__form-field">
-          <readonly-field
-            :label="'sponsorship-form.total-amount' | globalize"
-            :value="totalAmount"
-            :error-message="getFieldErrorMessage(
-              'totalAmount',
-              {
-                available: assetBalance.balance,
-              }
-            )"
-          />
-        </div>
-      </div>
-
-      <template v-if="formMixin.isConfirmationShown">
-        <form-confirmation
-          @cancel="hideConfirmation"
-          @ok="submit"
-          :is-pending="isFormSubmitting"
-          class="app__form-confirmation"
-        />
-      </template>
-
-      <template v-else>
-        <div class="app__form-actions">
-          <button
-            v-ripple
-            type="submit"
-            class="app__button-raised sponsorship-form__submit-btn"
-          >
-            {{ 'sponsorship-form.create-contract-btn' | globalize }}
-          </button>
-        </div>
-      </template>
-    </form>
+        <template v-else>
+          <div class="app__form-actions">
+            <button
+              v-ripple
+              type="submit"
+              :disabled="!isTotalAmountExists"
+              class="app__button-raised sponsorship-form__submit-btn"
+            >
+              {{ 'sponsorship-form.create-contract-btn' | globalize }}
+            </button>
+          </div>
+        </template>
+      </form>
+    </template>
+    <template v-else>
+      <no-data-message
+        icon-name="coins"
+        :title="'sponsorship-form.no-assets-msg-title' | globalize"
+        :message="'sponsorship-form.no-assets-msg-description' | globalize"
+      />
+    </template>
   </div>
 </template>
 
 <script>
 import FormMixin from '@/vue/mixins/form.mixin'
 import ReadonlyField from '@/vue/fields/ReadonlyField'
+import NoDataMessage from '@/vue/common/NoDataMessage'
 
 import { required, maxValueBig } from '@validators'
 import { mapGetters } from 'vuex'
@@ -101,6 +112,7 @@ export default {
   name: 'sponsorhip-form',
   components: {
     ReadonlyField,
+    NoDataMessage,
   },
 
   mixins: [FormMixin],
@@ -130,7 +142,7 @@ export default {
 
   computed: {
     ...mapGetters([
-      vuexTypes.balancesAssets,
+      vuexTypes.accountBalances,
       vuexTypes.accountBalanceByCode,
       vuexTypes.accountId,
     ]),
@@ -146,6 +158,16 @@ export default {
         MathUtil.roundingModes.ROUND_DOWN
       )
     },
+
+    balancesAssets () {
+      return this.accountBalances
+        .filter(i => +i.balance > 0)
+        .map(item => item.asset)
+    },
+
+    isTotalAmountExists () {
+      return Boolean(+this.totalAmount)
+    },
   },
 
   created () {
@@ -160,11 +182,11 @@ export default {
     async submit () {
       this.isFormSubmitting = true
       try {
-        const contract = this.buildContract()
-        await api.postWithSignature('/integrations/sponsorship/contracts', contract)
         const { data: paymentAccount } = await api.getWithSignature('/integrations/sponsorship/info')
         const paymentOperation = this.buildPaymentOperation(paymentAccount.id)
-        await api.postOperations(paymentOperation)
+        const paymentTx = await api.getTransaction(paymentOperation)
+        const contract = this.buildContract(paymentTx)
+        await api.postWithSignature('/integrations/sponsorship/contracts', contract)
 
         Bus.success('sponsorship-form.create-contract-successful')
         this.$emit(EVENTS.contractCreated)
@@ -175,7 +197,7 @@ export default {
       this.hideConfirmation()
     },
 
-    buildContract () {
+    buildContract (paymentTx) {
       return {
         consumer_business: this.businessAsset.owner,
         consumers_asset: this.businessAsset.asset.code,
@@ -183,6 +205,7 @@ export default {
         sponsor_business: this.accountId,
         sponsor_amount: this.form.amount,
         max_users_count: this.businessAsset.holders,
+        tx: paymentTx,
       }
     },