+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+## [1.4.0-rc.2] - 2019-11-14
+#### Fixed
+- A bub with clients export
+
 ## [1.4.0-rc.1] - 2019-11-14
 #### Added
 - Clients export
@@ -413,7 +417,8 @@ for further information about branching and tagging conventions.
 #### Removed
 - `RECOVERY_MODE` config key due to unused anymore
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.4.0-rc.1...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.4.0-rc.2...HEAD
+[1.4.0-rc.2]: https://github.com/tokend/web-client/compare/1.4.0-rc.1...1.4.0-rc.2
 [1.4.0-rc.1]: https://github.com/tokend/web-client/compare/1.4.0-rc.0...1.4.0-rc.1
 [1.4.0-rc.0]: https://github.com/tokend/web-client/compare/1.3.1...1.4.0-rc.0
 [1.3.1]: https://github.com/tokend/web-client/compare/1.3.0...1.3.1